Finished refactor, only base class extensions left in root
