+++ lib/controller/addressbookcontroller.php
@@ -10,7 +10,7 @@
 namespace OCA\Contacts\Controller;
 
 use OCA\Contacts\App,
-	OCA\Contacts\JSONResponse,
+	OCA\Contacts\Service\JSONResponse,
 	OCA\Contacts\Utils\JSONSerializer,
 	OCA\Contacts\Controller,
 	OCP\AppFramework\Http;
