+++ lib/controller/exportcontroller.php
@@ -11,9 +11,9 @@
 namespace OCA\Contacts\Controller;
 
 use OCA\Contacts\App,
-	OCA\Contacts\JSONResponse,
+	OCA\Contacts\Service\JSONResponse,
 	OCA\Contacts\Controller,
-	OCA\Contacts\TextDownloadResponse,
+	OCA\Contacts\Service\TextDownloadResponse,
 	Sabre\VObject;
 
 /**
