+++ lib/service/jsonresponse.php
@@ -8,7 +8,7 @@
  * See the COPYING-README file.
  */
 
-namespace OCA\Contacts;
+namespace OCA\Contacts\Service;
 use OCP\AppFramework\Http\JSONResponse as OriginalResponse,
 	OCP\AppFramework\Http;
