+++ settings.php
@@ -1,6 +1,6 @@
 <?php
 
 $tmpl = new OCP\Template( 'contacts', 'settings');
-$tmpl->assign('addressbooks', OCA\Contacts\Addressbook::all(OCP\USER::getUser()));
+$tmpl->assign('addressbooks', OCA\Contacts\Model\Addressbook::all(OCP\USER::getUser()));
 
 $tmpl->printPage();
